Display is 80% done
The logic for when numbers has to be entered or when an operation has to be chosen is implemented except for the brackets. Further work is required to integrate the logic of brackets into the display box.

Currently there is no function to remove the most recently added symbol/number.

Not sure if this is going to be the way we make the display, because I'm still not sure how the whole thing will be calculated.